requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
asgiref==3.5.2
    # via django
black==22.12.0
    # via -r requirements.in
build==0.9.0
    # via pip-tools
cfgv==3.3.1
    # via pre-commit
click==8.1.3
    # via
    #   black
    #   pip-tools
distlib==0.3.6
    # via virtualenv
django==4.1.4
    # via
    #   -r requirements.in
    #   django-extensions
    #   django-phonenumber-field
    #   djangorestframework
django-environ==0.9.0
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-phonenumber-field==7.0.1
    # via -r requirements.in
djangorestframework==3.14.0
    # via -r requirements.in
filelock==3.8.2
    # via virtualenv
flake8==6.0.0
    # via -r requirements.in
identify==2.5.10
    # via pre-commit
mccabe==0.7.0
    # via flake8
mypy-extensions==0.4.3
    # via black
nodeenv==1.7.0
    # via pre-commit
packaging==22.0
    # via build
pathspec==0.10.3
    # via black
pep517==0.13.0
    # via build
pip-tools==6.12.1
    # via -r requirements.in
platformdirs==2.6.0
    # via
    #   black
    #   virtualenv
pre-commit==2.20.0
    # via -r requirements.in
psycopg2==2.9.5
    # via -r requirements.in
pycodestyle==2.10.0
    # via flake8
pyflakes==3.0.1
    # via flake8
pytz==2022.6
    # via djangorestframework
pyyaml==6.0
    # via pre-commit
sqlparse==0.4.3
    # via django
toml==0.10.2
    # via pre-commit
tomli==2.0.1
    # via
    #   black
    #   build
virtualenv==20.17.1
    # via pre-commit
wheel==0.38.4
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
